Refactor task management: integrate useTasks hook, enhance AddTask and TaskBoard components, and implement TaskList for better task organization
