+++ frontEnd/src/hooks/useTasks.tsx
@@ -1,24 +1,29 @@
-import { useState, useEffect } from "react";
-import api from "../services/api";
+import { useState, useEffect } from 'react';
+import { Task } from '../types';
+import api from '../services/api';
 
-type Task = {
-    id: number
-    title: string
-    description: string
-    status: 'Pendente' | 'Em andamento' | 'Feito'
-}
+export const useTasks = () => {
+  const [tasks, setTasks] = useState<Task[]>([]);
 
-export function useTasks() {
-    const [tasks, setTasks] = useState<Task[]>([])
+  useEffect(() => {
+    const fetchTasks = async () => {
+      const response = await api.get('/');
+      setTasks(response.data);
+    };
+    fetchTasks();
+  }, []);
 
-    useEffect(() => {
-        api.get('/tasks').then(response => setTasks(response.data))
-    }, [])
+  const addTask = (newTask: Task) => {
+    setTasks((prevTasks) => [...prevTasks, newTask]);
+  };
 
-    const addTask = async (title: string, description: string) => {
-        const response = await api.post('/tasks', { title, description })
-        setTasks([...tasks, response.data])
-    }
+  const updateTask = (updatedTask: Task) => {
+    setTasks((prevTasks) =>
+      prevTasks.map((task) =>
+        task.id === updatedTask.id ? updatedTask : task
+      )
+    );
+  };
 
-    return { tasks, addTask }
-}
+  return { tasks, setTasks, addTask, updateTask };
+};
